requirements.txt
streamlit==1.32.0
snowflake-connector-python
pandas
yfinance
plotly
openpyxl
